test(app): withTemporaryDefaults를 공용 지원 파일로 분리
EventTapDecisionTests가 크로스 파일로 소비하는 필수 격리 헬퍼가 한 기능의
테스트 파일(SafetyToggleTests) 안에 숨어 있었다 — QwertyLayoutCondition
선례대로 독립 지원 파일로 이동 (기능 테스트 삭제·개명 시 무관한 테스트
컴파일이 깨지는 결합 제거).
